# Use fast HPACK comparisons when not checking sensitive headers