@normanmaurer 2019-06-20
@carl-mastrangelo do you have a benchmark by any chance?