@carl-mastrangelo 2019-06-20
The effect is not so pronounced, but it helps in the larger cases.  My guess why insensitive is slower for small headers is that there are multiple lookups involved in the dynamic and static table.  It only gets faster when the headers are already in the table.

```
BEFORE:

Benchmark                     (duplicates)  (limitToAscii)  (sensitive)  (size)  Mode  Cnt       Score       Error  Units
HpackEncoderBenchmark.encode          true            true         true   SMALL  avgt   10    2572.595 ±    16.184  ns/op
HpackEncoderBenchmark.encode          true            true         true  MEDIUM  avgt   10   19580.815 ±   397.780  ns/op
HpackEncoderBenchmark.encode          true            true         true   LARGE  avgt   10  379456.381 ±  2059.919  ns/op
HpackEncoderBenchmark.encode          true            true        false   SMALL  avgt   10     730.579 ±     8.116  ns/op
HpackEncoderBenchmark.encode          true            true        false  MEDIUM  avgt   10    2087.590 ±    84.644  ns/op
HpackEncoderBenchmark.encode          true            true        false   LARGE  avgt   10   11725.228 ±    89.298  ns/op
HpackEncoderBenchmark.encode          true           false         true   SMALL  avgt   10     555.971 ±     5.120  ns/op
HpackEncoderBenchmark.encode          true           false         true  MEDIUM  avgt   10    2831.874 ±    41.801  ns/op
HpackEncoderBenchmark.encode          true           false         true   LARGE  avgt   10   36054.025 ±   179.504  ns/op
HpackEncoderBenchmark.encode          true           false        false   SMALL  avgt   10     340.337 ±     3.313  ns/op
HpackEncoderBenchmark.encode          true           false        false  MEDIUM  avgt   10    1006.817 ±     8.942  ns/op
HpackEncoderBenchmark.encode          true           false        false   LARGE  avgt   10    8784.168 ±   164.014  ns/op
HpackEncoderBenchmark.encode         false            true         true   SMALL  avgt   10    2561.934 ±    27.056  ns/op
HpackEncoderBenchmark.encode         false            true         true  MEDIUM  avgt   10   22061.105 ±   154.533  ns/op
HpackEncoderBenchmark.encode         false            true         true   LARGE  avgt   10  435744.897 ±  8853.388  ns/op
HpackEncoderBenchmark.encode         false            true        false   SMALL  avgt   10    2737.683 ±    47.142  ns/op
HpackEncoderBenchmark.encode         false            true        false  MEDIUM  avgt   10   22385.146 ±    98.430  ns/op
HpackEncoderBenchmark.encode         false            true        false   LARGE  avgt   10  408159.698 ± 12044.931  ns/op
HpackEncoderBenchmark.encode         false           false         true   SMALL  avgt   10     544.213 ±     3.279  ns/op
HpackEncoderBenchmark.encode         false           false         true  MEDIUM  avgt   10    2908.978 ±    31.026  ns/op
HpackEncoderBenchmark.encode         false           false         true   LARGE  avgt   10   36471.262 ±  1044.010  ns/op
HpackEncoderBenchmark.encode         false           false        false   SMALL  avgt   10     609.305 ±     4.371  ns/op
HpackEncoderBenchmark.encode         false           false        false  MEDIUM  avgt   10    3223.946 ±    23.505  ns/op
HpackEncoderBenchmark.encode         false           false        false   LARGE  avgt   10   39975.152 ±   655.196  ns/op

AFTER:

Benchmark                     (duplicates)  (limitToAscii)  (sensitive)  (size)  Mode  Cnt       Score      Error  Units
HpackEncoderBenchmark.encode          true            true         true   SMALL  avgt   10    2580.732 ±   39.935  ns/op
HpackEncoderBenchmark.encode          true            true         true  MEDIUM  avgt   10   19963.569 ±  417.324  ns/op
HpackEncoderBenchmark.encode          true            true         true   LARGE  avgt   10  361053.507 ± 2141.450  ns/op
HpackEncoderBenchmark.encode          true            true        false   SMALL  avgt   10     774.761 ±    3.446  ns/op
HpackEncoderBenchmark.encode          true            true        false  MEDIUM  avgt   10    1601.366 ±   11.409  ns/op
HpackEncoderBenchmark.encode          true            true        false   LARGE  avgt   10    4852.643 ±   43.546  ns/op
HpackEncoderBenchmark.encode          true           false         true   SMALL  avgt   10     551.301 ±    3.420  ns/op
HpackEncoderBenchmark.encode          true           false         true  MEDIUM  avgt   10    2821.285 ±   11.290  ns/op
HpackEncoderBenchmark.encode          true           false         true   LARGE  avgt   10   35519.282 ±  194.778  ns/op
HpackEncoderBenchmark.encode          true           false        false   SMALL  avgt   10     251.308 ±    2.342  ns/op
HpackEncoderBenchmark.encode          true           false        false  MEDIUM  avgt   10     509.551 ±    3.531  ns/op
HpackEncoderBenchmark.encode          true           false        false   LARGE  avgt   10    1648.384 ±   17.924  ns/op
HpackEncoderBenchmark.encode         false            true         true   SMALL  avgt   10    2782.013 ±  151.077  ns/op
HpackEncoderBenchmark.encode         false            true         true  MEDIUM  avgt   10   21969.109 ±   63.417  ns/op
HpackEncoderBenchmark.encode         false            true         true   LARGE  avgt   10  415850.888 ± 1507.624  ns/op
HpackEncoderBenchmark.encode         false            true        false   SMALL  avgt   10    2699.697 ±   34.675  ns/op
HpackEncoderBenchmark.encode         false            true        false  MEDIUM  avgt   10   23074.582 ±  105.401  ns/op
HpackEncoderBenchmark.encode         false            true        false   LARGE  avgt   10  405572.113 ± 2954.830  ns/op
HpackEncoderBenchmark.encode         false           false         true   SMALL  avgt   10     543.590 ±    2.325  ns/op
HpackEncoderBenchmark.encode         false           false         true  MEDIUM  avgt   10    2943.186 ±   12.283  ns/op
HpackEncoderBenchmark.encode         false           false         true   LARGE  avgt   10   36131.091 ±  119.945  ns/op
HpackEncoderBenchmark.encode         false           false        false   SMALL  avgt   10     601.432 ±    5.279  ns/op
HpackEncoderBenchmark.encode         false           false        false  MEDIUM  avgt   10    3225.028 ±   46.043  ns/op
HpackEncoderBenchmark.encode         false           false        false   LARGE  avgt   10   39672.871 ±  585.702  ns/op
```