@carl-mastrangelo 2019-06-20
@normanmaurer I am going to send a followup PR.  Disabling the huffman coder makes this ridiculously faster, to the point where this PR makes only a minor difference.