@normanmaurer 2019-06-20
@carl-mastrangelo so should we close this then ?