@carl-mastrangelo 2019-06-20
@normanmaurer I think it will still have a minor benefit for large header values, just small headers won't see it.   The two changes are independent, but the other is more important.