@normanmaurer 2019-06-20
+1

> Am 20.06.2019 um 11:03 schrieb Carl Mastrangelo <notifications@github.com>:
> 
> @carl-mastrangelo commented on this pull request.
> 
> In codec-http2/src/main/java/io/netty/handler/codec/http2/HpackHeaderField.java:
> 
> > @@ -59,21 +60,16 @@ final int size() {
>  
>      @Override
>      public final int hashCode() {
> -        // TODO(nmittler): Netty's build rules require this. Probably need a better implementation.
> -        return super.hashCode();
> +        throw new UnsupportedOperationException();
> I can remove hashCode and equals. WDYT?
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.